// src/services/Fivem/proto/master.d.ts
export namespace master {
// ...
  interface IServerData {
// ...
    vars?: ({ [k: string]: string } | null);

    /** ServerData enhancedHostSupport */
    enhancedHostSupport?: (boolean | null);

    /** ServerData upvotePower */
    upvotePower?: (number | null);

    /** ServerData burstPower */
    burstPower?: (number | null);

    /** ServerData connectEndPoints */
    connectEndPoints?: (string[] | null);
  }
// ...
}